Support slice lengths to be encoded as an u32
Slice lengths (an usize) are encoded in bincode as a u64 value. Given
that the width of usize is platform-specific, an usize value encoded
on a platform with a wider usize has the potential for decoding to
fail on a platform with a narrower usize. If we know the usize width
of the target decoding platform at encoding time, it could be
preferable to fail at encoding time instead, giving us the possibility
to detect this error.

This commit introduces the ability to choose, at encoding time,
between slices being 32 or 64 bits wide. These widths are chosen for
being the two most common slice length type widths on common
architectures. If a slice length value is given that is out of range
for the chosen slice length type, an error will happen at encoding
time.