Add EncodedSize trait to calculate encoded sizes
For some storage back-ends it's necessary to know the size of the buffer to pre-allocate before encoding.

The bincode 1.0 version had `bincode::serialized_size`, which would tell you the serialized size of data.

We add a new `EncodedSize` trait which accomplishes the same objective, and implement that for all types.  Implementations for custom structs and enums can be derived using `#[derive(bincode::EncodedSize)]`.